requirements.txt
flask
pycrypto
gnupg
sysv_ipc
tabulate
gevent
mysqldb